Make easier for users to point to their geonetwork application
If a new user executes the suggested command
`./gemet-to-simpleskos.sh en fr de nl it` without updating the value of
`GEONETWORK_HOME` inside `gemet-to-simpleskos.sh` it will fail with a
confusing error:

```bash
Creating thesaurus ...
Error: Could not find or load main class net.sf.saxon.Transform
```

This happens because the CLASSPATH for the java call should contain
`$GEONETWORK_HOME/WEB-INF/lib/saxon-9.1.0.8b-patch.jar` and the
`GEONETWORK_HOME` is set inside `gemet-to-simpleskos.sh` to a default
path. If geonetwork is not present at that path the script will fail.

A better approach is make the user set the `GEONETWORK_HOME` variable
before calling the script.

I also added a check to warn the user if the script could not locate
the saxon jar. In this case it will print an error message and exit
with an error code.